refactor: move source types and utils to folders, added new source
