+++ services/ingestion/source_types/rss.py
@@ -50,7 +50,7 @@ async def fetch_rss(source: dict) -> list[dict]:
         response = await client.get(
             source["url"],
             # Mimic a browser user agent — some feeds block default httpx UA
-            headers={"User-Agent": "Mozilla/5.0"}
+            headers={"User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"}
         )
         response.raise_for_status()
